moving dummy passes to a different file
