+++ test/python/transpiler/test_pass_scheduler.py
@@ -14,144 +14,11 @@ import logging
 
 from qiskit import QuantumRegister, QuantumCircuit
 from qiskit.dagcircuit import DAGCircuit
-from qiskit.transpiler import PassManager, transpile, TransformationPass, AnalysisPass, \
-    TranspilerAccessError
+from qiskit.transpiler import PassManager, transpile, TranspilerAccessError
+from ._dummy_passes import *
 from ..common import QiskitTestCase
 
 logger = "LocalLogger"
-
-
-class DummyTP(TransformationPass):
-    """ A dummy transformation pass."""
-
-    def run(self, dag, property_set):
-        logging.getLogger(logger).info('run transformation pass %s', self.name)
-
-
-class DummyAP(AnalysisPass):
-    """ A dummy analysis pass."""
-
-    def run(self, dag, property_set):
-        logging.getLogger(logger).info('run analysis pass %s', self.name)
-
-
-class PassA_TP_NR_NP(DummyTP):
-    """ A dummy pass without any requires/preserves.
-    TP: Transformation Pass
-    NR: No requires
-    NP: No preserves
-    """
-    pass
-
-
-class PassB_TP_RA_PA(DummyTP):
-    """ A dummy pass that requires PassA_TP_NR_NP and preserves it.
-    TP: Transformation Pass
-    RA: Requires PassA
-    PA: Preserves PassA
-    """
-    requires = [PassA_TP_NR_NP()]
-    preserves = [PassA_TP_NR_NP()]
-
-
-class PassC_TP_RA_PA(DummyTP):
-    """ A dummy pass that requires PassA_TP_NR_NP and preserves it.
-    TP: Transformation Pass
-    RA: Requires PassA
-    PA: Preserves PassA
-    """
-    requires = [PassA_TP_NR_NP()]
-    preserves = [PassA_TP_NR_NP()]
-
-
-class PassD_TP_NR_NP(DummyTP):
-    """ A dummy transfomation pass that takes an argument.
-    TP: Transformation Pass
-    NR: No Requires
-    NP: No Preserves
-    """
-
-    def __init__(self, argument1):
-        super().__init__()
-        self.argument1 = argument1
-
-    def run(self, dag, property_set):
-        super().run(dag, property_set)
-        logging.getLogger(logger).info('argument %s', self.argument1)
-
-
-class PassE_AP_NR_NP(DummyAP):
-    """ A dummy analysis pass that takes an argument.
-    AP: Analysis Pass
-    NR: No Requires
-    NP: No Preserves
-    """
-
-    def __init__(self, argument1):
-        super().__init__()
-        self.argument1 = argument1
-
-    def run(self, dag, property_set):
-        super().run(dag, property_set)
-        property_set['property'] = self.argument1
-        logging.getLogger(logger).info('set property as %s', property_set['property'])
-
-
-class PassF_reduce_dag_property(DummyTP):
-    """ A dummy transformation pass that (sets and)reduces a property in the DAG.
-    AP: Analysis Pass
-    NR: No Requires
-    NP: No Preserves
-    """
-
-    def run(self, dag, property_set):
-        super().run(dag, property_set)
-        if not hasattr(dag, 'property'):
-            dag.property = 8
-        dag.property = round(dag.property * 0.8)
-        logging.getLogger(logger).info('dag property = %i', dag.property)
-
-
-class PassG_calculates_dag_property(DummyAP):
-    """ A dummy transformation pass that (sets and)reduces a property in the DAG.
-    AP: Analysis Pass
-    NR: No Requires
-    NP: No Preserves
-    """
-
-    def run(self, dag, property_set):
-        super().run(dag, property_set)
-        prop = dag.property
-        property_set['property'] = prop
-        logging.getLogger(logger).info('set property as %s (from dag.property)',
-                                       property_set['property'])
-
-
-class PassH_Bad_TP(DummyTP):
-    """ A dummy transformation pass tries to modify the property set.
-    NR: No Requires
-    NP: No Preserves
-    """
-
-    def run(self, dag, property_set):
-        super().run(dag, property_set)
-        property_set['property'] = "value"
-        logging.getLogger(logger).info('set property as %s', property_set['property'])
-
-
-class PassI_Bad_AP(DummyAP):
-    """ A dummy analysis pass tries to modify the dag.
-    NR: No Requires
-    NP: No Preserves
-    """
-
-    def run(self, dag, property_set):
-        super().run(dag, property_set)
-        cx_runs = dag.collect_runs(["cx"])
-        logging.getLogger(logger).info('cx_runs: %s', cx_runs)
-        dag._remove_op_node(cx_runs.pop()[0])
-        logging.getLogger(logger).info('done removing')
-
 
 class TestUseCases(QiskitTestCase):
     """ The pass manager schedules passes in, sometimes, tricky ways. These tests combine passes in
